+++ src/components/App/filters/filters.js
@@ -20,10 +20,10 @@ export function cropAddress (value) {
 
 export function roleIdToString (roleId) {
   return {
-    [config.ACCOUNT_ROLES.notVerified]: 'Unverified',
-    [config.ACCOUNT_ROLES.general]: 'Individual',
-    [config.ACCOUNT_ROLES.corporate]: 'Corporate'
-  }[roleId]
+    [+config.ACCOUNT_ROLES.notVerified]: 'Unverified',
+    [+config.ACCOUNT_ROLES.general]: 'Individual',
+    [+config.ACCOUNT_ROLES.corporate]: 'Corporate'
+  }[+roleId]
 }
 
 export function lowerCase (string) {